Resolved various issues from JIRA
- Updated VDK to 28974 for vdkConvert improvements
- Tidied up how WKT is loaded and displayed (Indirectly Resolves EVA-340)
- Further improvements to the convert system
  - Added the cancel system from VDK (Resolves EVA-295)
  - Fixed an issue where the convert jobs would be marked completed when no licenses were available (Resolves EVA-338)
  - Improved the error messages so it says if a convert job failed, succeeded was cancelled or is pending for license reasons
  - Images dropped onto client are now assumed to be watermarks for convert (Resolves EVA-341)

+++ src/vcGIS.h
@@ -4,7 +4,7 @@
 #include "udPlatform/udMath.h"
 #include "udPlatform/udGeoZone.h"
 
-typedef uint16_t vcSRID;
+typedef int32_t vcSRID;
 
 struct vcGISSpace
 {
